fixed errors w/includes, changed base email footer, added new email footers for error
