+++ includes/languages/english/modules/payment/bitpay.php
@@ -22,7 +22,7 @@
   // Text Messages
   define('MODULE_PAYMENT_BITPAY_TEXT_TITLE', 'Bitcoin via BitPay');
   define('MODULE_PAYMENT_BITPAY_TEXT_DESCRIPTION', 'Use bitpay.com\'s invoice processing server to automatically accept bitcoins.');
-  define('MODULE_PAYMENT_BITPAY_TEXT_EMAIL_FOOTER', 'You just paid with bitcoins via bitpay.com -- Thanks!');
+  define('MODULE_PAYMENT_BITPAY_TEXT_EMAIL_FOOTER', 'You will receive further emails from bitpay.com when the status of your order is updated.');
 
   // Error Messages
   define('MODULE_PAYMENT_BITPAY_BAD_CURRENCY', 'Currency not supported by bitpay.com.  Please choose another currency.');
